% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cu_metrics.R
\name{calc_cu_metrics}
\alias{calc_cu_metrics}
\title{Channel Unit Metrics - Site Scale}
\usage{
calc_cu_metrics(folder_name = NULL)
}
\arguments{
\item{folder_name}{character string of the folder name where Survey123 data has been saved}
}
\description{
Combine all the channel unit measurements for a particular site and calculate channel unit metrics
}
\author{
Kevin See
}
