% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_names.R
\name{check_col_names}
\alias{check_col_names}
\title{QC - Check Column Names}
\usage{
check_col_names(
  qc_df = NULL,
  otg_type = c("surveyPoint_0.csv", "CU_1.csv", "Wood_2.csv", "Jam_3.csv",
    "Undercut_4.csv", "Discharge_5.csv", "DischargeMeasurements_6.csv")
)
}
\arguments{
\item{qc_df}{The survey data frame to be QC'd}

\item{otg_type}{what type of on-the-ground (OTG) data would the user like to deal with? As of
the initiation of the function a character vector of the file names e.g., "surveyPoint_0.csv",
"CU_1.csv", etc.}
}
\value{
a message
}
\description{
Compare columns of a tibble or data frame to the expected
columns defined in \code{get_otg_col_specs()}
}
\author{
Mike Ackerman
}
