% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na.R
\name{check_na}
\alias{check_na}
\title{QC - Check For NAs}
\usage{
check_na(qc_df = NULL, cols_to_check = NULL)
}
\arguments{
\item{qc_df}{The survey data frame to be QC'd}
}
\value{
\code{NULL} or a tibble from \code{qc_tbl()}
}
\description{
Check for unexpected or important NAs within
specified columns.
}
\author{
Mike Ackerman
}
