% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_files.R
\name{compare_files}
\alias{compare_files}
\title{Compare Files and Find Differences}
\usage{
compare_files(file1 = NA, file2 = NA, ...)
}
\arguments{
\item{file1}{a path to the first file of interest, including file extension, perhaps the raw untouched data.}

\item{file2}{a path to the second file of interest, including file extension, perhaps the data files that have been
modified or fixed by the user, presumably to remedy errors identified during data import or
QA/QC.}
}
\value{
fill out return when ready
}
\description{
Compare two files, each presumably containing the same data, except with
one folder containing files that have been modified or "fixed" by the user. For the \code{DASH}
R package, the function can be useful for comparing a file containing raw untouched
on-the-ground (OTG) data from Survey123 with a file containing the same data except modified
by the user, presumably in an attempt to remedy errors identified during data import or
QA/QC. Both \code{file1} and \code{file2} should only contain entire paths, including file name and extension, each representing
a survey containing the otg data. These files must be either .csv or .xlsx/.xls files.
}
\author{
Kevin See
}
