% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_folders.R
\name{compare_folders}
\alias{compare_folders}
\title{Compare Folders and Find Differences}
\usage{
compare_folders(path1 = ".", path2 = ".")
}
\arguments{
\item{path1}{a path to the first folder of interest, perhaps the raw untouched data. \code{path1}
should only contain folders (no files), and each of those folders should only contain files (no folders).}

\item{path2}{a path to the second folder of interest, perhaps the data files that have been
modified or fixed by the user, presumably to remedy errors identified during data import or
QA/QC.}
}
\value{
fill out return when ready
}
\description{
Compare two folders, each presumably containing the same files, except with
one folder containing files that have been modified or "fixed" by the user. For the \code{DASH}
R package, the function can be useful for comparing a folder containing raw untouched
on-the-ground (OTG) data from Survey123 with a folder containing the same files except modified
by the user, presumably in an attempt to remedy errors identified during data import or
QA/QC. Both \code{path1} and \code{path2} should only contain folders (no files) each representing
a survey, and each of those folders should only contain files (no folders) containing the otg
data.
}
\author{
Mike Ackerman
}
