% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_cu_metrics.R
\name{compile_cu_metrics}
\alias{compile_cu_metrics}
\title{Channel Unit Metrics - All Sites}
\usage{
compile_cu_metrics(folder_name = NULL)
}
\arguments{
\item{folder_name}{character string of the folder name where list of folders with Survey123 data}
}
\description{
Combine all the channel unit measurements for a particular site and calculate channel unit metrics
}
\author{
Kevin See
}
