% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileQAQC.R
\name{compileQAQC}
\alias{compileQAQC}
\title{Compile QAQC errors}
\usage{
compileQAQC(folder_name = NULL)
}
\arguments{
\item{folder_name}{character string of the folder name where list of folders with Survey123 data}
}
\description{
Combine all of the QAQC errors from the dashQAQC() function across files within a directory
}
\author{
Mike Ackerman
}
