% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_tbl.R
\name{create_empty_tbl}
\alias{create_empty_tbl}
\title{Create Empty Data Tables}
\usage{
create_empty_tbl(
  otg_type = c("surveyPoint_0.csv", "CU_1.csv", "Wood_2.csv", "Jam_3.csv",
    "Undercut_4.csv", "Discharge_5.csv", "DischargeMeasurements_6.csv")
)
}
\arguments{
\item{otg_type}{what type of on-the-ground (OTG) data would the user like to deal with?}
}
\value{
an empty tibble
}
\description{
Create a empty tibble with the same column names and
column types as from a typical survey of the specified type.
}
\author{
Kevin See
}
