% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CU_1_UpperLemhi1_2019.R
\docType{data}
\name{CU_1_UpperLemhi1_2019}
\alias{CU_1_UpperLemhi1_2019}
\title{Channel Unit (CU) data from Upper Lemhi 1 survey 2019}
\format{Initially exported from Survey123 as a csv file}
\usage{
CU_1_UpperLemhi1_2019
}
\description{
This is an example dataset for \code{DASH()} for channel units from the Upper Lemhi 1 survey in 2019.
}
\keyword{DASH}
\keyword{Lemhi}
\keyword{channel}
\keyword{unit}
