% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashQAQC.R
\name{dash_QAQC}
\alias{dash_QAQC}
\title{DASH Data Quality Control}
\usage{
dash_QAQC(folder_name = NULL)
}
\arguments{
\item{folder_name}{character string of the folder name where Survey123 data has been saved}
}
\description{
Read DASH data and perform a quick data quality control and validation
}
\author{
Mike Ackerman
}
