% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DischargeLocations_5_Hawley2019.R
\docType{data}
\name{DischargeLocations_5_Hawley2019}
\alias{DischargeLocations_5_Hawley2019}
\title{Location metadata for discharge measurements from Hawley survey 2019}
\format{Initially exported from Survey123 as a csv file}
\usage{
DischargeLocations_5_Hawley2019
}
\description{
This is an example dataset for \code{DASH()} and for discharge measurement locations from the Hawley Creek survey in 2019.
}
\keyword{DASH}
\keyword{Hawley}
\keyword{discharge}
\keyword{location}
