% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DischargeLocations_5_UpperLemhi1_2019.R
\docType{data}
\name{DischargeLocations_5_UpperLemhi1_2019}
\alias{DischargeLocations_5_UpperLemhi1_2019}
\title{Location metadata for discharge measurements from Upper Lemhi 1 survey 2019}
\format{Initially exported from Survey123 as a csv file}
\usage{
DischargeLocations_5_UpperLemhi1_2019
}
\description{
This is an example dataset for \code{DASH()} and for discharge measurement locations from the Upper Lemhi 1 survey in 2019.
}
\keyword{DASH}
\keyword{Lemhi}
\keyword{discharge}
\keyword{location}
