% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DischargeMeasurements_6_Hawley2019.R
\docType{data}
\name{DischargeMeasurements_6_Hawley2019}
\alias{DischargeMeasurements_6_Hawley2019}
\title{Discharge measurement data from Hawley survey 2019}
\format{Initially exported from Survey123 as a csv file}
\usage{
DischargeMeasurements_6_Hawley2019
}
\description{
This is an example dataset for \code{DASH()} and for discharge measurements from the Hawley Creek survey in 2019.
}
\keyword{DASH}
\keyword{Hawley}
\keyword{discharge}
