% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DischargeMeasurements_6_UpperLemhi1_2019.R
\docType{data}
\name{DischargeMeasurements_6_UpperLemhi1_2019}
\alias{DischargeMeasurements_6_UpperLemhi1_2019}
\title{Discharge measurement data from Upper Lemhi 1 survey 2019}
\format{Initially exported from Survey123 as a csv file}
\usage{
DischargeMeasurements_6_UpperLemhi1_2019
}
\description{
This is an example dataset for \code{DASH()} and for discharge measurements from the Upper Lemhi 1 survey in 2019.
}
\keyword{DASH}
\keyword{Lemhi}
\keyword{discharge}
