% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_fish_cover.R
\name{fix_fish_cover}
\alias{fix_fish_cover}
\title{Fix Fish Cover Columns}
\usage{
fix_fish_cover(
  cu_df = NULL,
  cover_cols = c("Overhanging Cover", "Aquatic Vegetation", "Woody Debris Cover",
    "Artificial Cover", "Total No Cover")
)
}
\arguments{
\item{cu_df}{data.frame of \code{otg_type =} "CU_1.csv" data that should contain
the \code{cover_cols} to be fixed.}

\item{cover_cols}{character vector of fish cover column names}
}
\value{
a tibble with some fish cover estimates resolved
}
\description{
Fixes typical issues identified in the fish cover columns of cu_df including
re-scaling estimates entered as proportions into percentages and re-scaling any estimates that
sum to 90 (likely due to a math error in the field) instead to 100. Any remaining issues (e.g.,
values sum outside the "expected" 100 - 130 or estimates are a combination of proportions and
percentages) should be identified by the \code{qc_cu()} function.
}
\author{
Mike Ackerman and Kevin See
}
