% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_nms.R
\name{get_file_nms}
\alias{get_file_nms}
\title{Get File Names}
\usage{
get_file_nms(path = ".")
}
\arguments{
\item{path}{a path to the directory containing the folders of interest. \code{path} should only contain
folders (no files), and each of those folders should only contain files (no folders).}
}
\value{
a data frame of all files contained in \code{path}
}
\description{
Create a data frame of folders and file names within \code{path}
}
\author{
Mike Ackerman and Kevin See
}
