% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_values.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute Missing Values}
\usage{
impute_missing_values(
  data_df = NULL,
  col_nm_vec = NULL,
  method = c("randomForestSRC", "missForest", "Hmisc"),
  ntree = 1000,
  nk = 4,
  ...
)
}
\arguments{
\item{data_df}{data.frame with potential missing values}

\item{col_nm_vec}{character vector of column names that should be either imputed and/or used to impute other columns}

\item{method}{which method to use, either by creating random forests using the \code{missForest} package, or by predictive mean matching using the \code{aregImpute()} function in the \code{Hmisc} package}

\item{ntree}{how many trees to build if \code{missForest} or \code{randomForestSRC}}

\item{nk}{how many knots to use in smoothing splines if using \code{method = 'Hmisc'}}

\item{...}{other arguments to be passed to either \code{missForest::missForest()}, \code{Hmisc::aregImpute()}, or \code{randomForestSRC::impute()} functions}
}
\value{
a data.frame with the same dimensions as the original, but with no missing values
}
\description{
Impute any missing values from selected columns within a data.frame
}
\author{
Kevin see
}
