% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{init_qc}
\alias{init_qc}
\title{Data: A tibble of initial QC results}
\format{
A tibble with 44 rows and 4 variables:
\describe{
\item{source}{the \code{otg_type} where the potential error was identified}
\item{path_nm}{the path name to the file where the potential error was identified}
\item{GlobalID}{the \strong{GlobalID} indicating the row where the potential error was identified}
\item{error_message}{a message indicating the nature of the potential error}
\item{location_id}{information on the site name, segment, and channel unit number were error was identified}
}
}
\usage{
init_qc
}
\description{
A tibble of initial QC results for the example dataset \code{otg_raw}
}
\keyword{datasets}
