% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jam_3_Hawley2019.R
\docType{data}
\name{Jam_3_Hawley2019}
\alias{Jam_3_Hawley2019}
\title{Wood jam data from Hawley survey 2019}
\format{Initially exported from Survey123 as a csv file}
\usage{
Jam_3_Hawley2019
}
\description{
This is an example dataset for \code{DASH()} and for wood jams from the Hawley Creek survey in 2019.
}
\keyword{DASH}
\keyword{Hawley}
\keyword{jam}
\keyword{wood}
