% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamMetrics.R
\name{jamMetrics}
\alias{jamMetrics}
\title{summarize the wood jam metrics for a DASH reach}
\usage{
jamMetrics(input = NULL)
}
\arguments{
\item{input}{the name of the tbl_df containing wood jam data for a DASH reach.}
}
\description{
summarize the wood jam metrics for a DASH reach
}
\author{
Richie Carmichael
}
