% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otg_to_cu.R
\name{otg_to_cu}
\alias{otg_to_cu}
\title{Rollup Channel Unit Data}
\usage{
otg_to_cu(
  survey_df = NULL,
  cu_df = NULL,
  wood_df = NULL,
  jam_df = NULL,
  undercut_df = NULL,
  discharge_df = NULL,
  fix_nas = TRUE,
  wood_impute_cols = c("length_m", "diameter_m"),
  jam_impute_cols = c("length_m", "width_m", "height_m", "estimated_number_of_pieces"),
  undercut_impute_cols = c("length_m", "width_25_percent_m", "width_50_percent_m",
    "width_75_percent_m"),
  ...
)
}
\arguments{
\item{survey_df}{data.frame of \verb{otg_type =} "surveyPoint_0.csv" containing information for
each site/survey}

\item{cu_df}{data.frame of \verb{otg_type =} "CU_1.csv" containing data for each channel unit}

\item{wood_df}{data.frame of \verb{otg_type =} "Wood_2.csv" containing the individual
wood data to be summarized by channel units}

\item{jam_df}{data.frame of \verb{otg_type =} "Jam_3.csv containing the individual
jam data to be summarized (rolled up) to the channel unit scale}

\item{undercut_df}{data.frame of \verb{otg_type =} "Undercut_4.csv" containing the individual
undercut data to be summarized (rolled up) to the channel unit scale}

\item{discharge_df}{input data.frame of \verb{otg_type =} "DischargeMeasurements_6.csv"
containing the station width, depth, and velocity data to estimate discharge}

\item{fix_nas}{if any of the length, width, height, or estimated number of pieces measurements
for an individual jam is missing i.e.,\code{NA}, would you like to fill them in? Default is \code{TRUE}, in which
case the \code{NA} values will be imputed using function \code{impute_missing_values()}}

\item{...}{other arguments to \code{impute_missing_values()}}
}
\value{
a data.frame summarizing data for channel units
}
\description{
Clean channel unit data (e.g., from \verb{otg_type =} "CU_1.csv") and join
survey/site information to it
}
\author{
Mike Ackerman, Kevin See, Richie Carmichael
}
