% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pebbleCountMetrics.R
\name{pebbleCountMetrics}
\alias{pebbleCountMetrics}
\title{summarize the pebble count metrics for a DASH reach}
\usage{
pebbleCountMetrics(input = NULL)
}
\arguments{
\item{input}{the name of the tbl_df containing pebble count data for a DASH reach.}
}
\description{
summarize the pebble count metrics for a DASH reach
}
\author{
Richie Carmichael
}
