% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc_tbl.R
\name{qaqc_tbl}
\alias{qaqc_tbl}
\title{Initiate Blank QA/QC Results Tibble}
\usage{
qaqc_tbl()
}
\value{
a blank tibble
}
\description{
Create a blank tibble to store messages from QA/QC of on-the-ground
collected data
}
\author{
Mike Ackerman
}
