% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_centerline.R
\name{qc_centerline}
\alias{qc_centerline}
\title{Quality Control - Centerlines}
\usage{
qc_centerline(
  cl_sf = NULL,
  data_id = "object_id",
  cols_to_check_nas = c("site_name", "year", "seg_num", "cu_num", "cu_type", "hab_rch")
)
}
\arguments{
\item{cl_sf}{The centerline \code{sf} object}

\item{cols_to_check_nas}{columns to check for NAs in}
}
\value{
a tibble with QC results
}
\description{
Quality control centerline shapefiles.
}
\author{
Kevin See
}
