% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_jam.R
\name{qc_jam}
\alias{qc_jam}
\title{Quality Control - Jams}
\usage{
qc_jam(
  qc_df = NULL,
  cols_to_check_nas = c("Length (m)", "Width (m)", "Height (m)",
    "Estimated Number of Pieces")
)
}
\arguments{
\item{qc_df}{The survey data frame to be QC'd}

\item{cols_to_check_nas}{columns to check for NAs in}
}
\value{
a tibble with QC results
}
\description{
Quality control jam data (e.g., from Jam_3.csv files) imported
using \code{read_otg_csv()} or \code{read_otg_csv_wrapper()}.
}
\author{
Kevin See
}
