% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_rollup.R
\name{qc_rollup}
\alias{qc_rollup}
\title{Quality Control - Channel Unit Rollup}
\usage{
qc_rollup(
  survey_df = NULL,
  cu_df = NULL,
  jam_df = NULL,
  undercut_df = NULL,
  wood_df = NULL,
  discharge_df = NULL
)
}
\arguments{
\item{survey_df}{data.frame of \verb{otg_type =} "surveyPoint_0.csv" containing information for
each site/survey}

\item{cu_df}{data.frame of \verb{otg_type =} "CU_1.csv" containing data for each channel unit}

\item{jam_df}{data.frame of \verb{otg_type =} "Jam_3.csv containing the individual
jam data to be summarized (rolled up) to the channel unit scale}

\item{undercut_df}{data.frame of \verb{otg_type =} "Undercut_4.csv" containing the individual
undercut data to be summarized (rolled up) to the channel unit scale}

\item{wood_df}{data.frame of \verb{otg_type =} "Wood_2.csv" containing the individual
wood data to be summarized by channel units}

\item{discharge_df}{input data.frame of \verb{otg_type =} "DischargeMeasurements_6.csv"
containing the station width, depth, and velocity data to estimate discharge}
}
\value{
a tibble with QC results
}
\description{
Quality control the channel unit rollup
performed by \code{otg_to_cu()}. This is primarily concerned with
making sure that all pieces of data are connected through the
parent global IDs.
}
\author{
Kevin See
}
