% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_survey.R
\name{qc_survey}
\alias{qc_survey}
\title{Quality Control - Survey Data}
\usage{
qc_survey(qc_df = NULL, otg_type = "surveyPoint_0.csv")
}
\arguments{
\item{qc_df}{The survey data frame to be QC'd}

\item{otg_type}{what type of on-the-ground (otg) data would the user like to deal with? As of the iniitation of the
function a character vector of the file name e.g., "surveyPoint_0.csv", "CU_1.csv", etc.}
}
\value{
a tibble with QC results
}
\description{
Quality control survey data (e.g., from surveyPoint_0.csv files) imported
using \code{read_otg_csv()} or \code{read_otg_csv_wrapper()}.
}
\author{
Mike Ackerman
}
