% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_tbl.R
\name{qc_tbl}
\alias{qc_tbl}
\title{Initiate Blank QC Results Tibble}
\usage{
qc_tbl(data_id = "GlobalID")
}
\arguments{
\item{id}{name of identifying column}
}
\value{
a blank tibble
}
\description{
Create a blank tibble to store messages from QA/QC of on-the-ground
collected data
}
\author{
Mike Ackerman
}
