% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_undercut.R
\name{qc_undercut}
\alias{qc_undercut}
\title{Quality Control - Undercuts}
\usage{
qc_undercut(
  qc_df = NULL,
  cols_to_check_nas = c("Location", "Length (m)", "Width 25\% (m)", "Width 50\% (m)",
    "Width 75\% (m)")
)
}
\arguments{
\item{qc_df}{The survey data frame to be QC'd}

\item{cols_to_check_nas}{columns to check for NAs in}
}
\value{
a tibble with QC results
}
\description{
Quality control undercut data (e.g., from Undercut_4.csv files) imported
using \code{read_otg_csv()} or \code{read_otg_csv_wrapper()}.
}
\author{
Kevin See
}
