% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_wood.R
\name{qc_wood}
\alias{qc_wood}
\title{Quality Control - Wood}
\usage{
qc_wood(
  qc_df = NULL,
  cols_to_check_nas = c("GlobalID", "Length (m)", "Diameter (m)", "Wet?",
    "Channel Forming?", "Ballasted?", "ParentGlobalID")
)
}
\arguments{
\item{qc_df}{The survey data frame to be QC'd}

\item{cols_to_check_nas}{columns to check for NAs in}
}
\value{
a tibble with QC results
}
\description{
Quality control wood data (e.g., from Wood_2.csv files) imported
using \code{read_otg_csv()} or \code{read_otg_csv_wrapper()}.
}
\author{
Kevin See
}
