% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_wrapper.R
\name{qc_wrapper}
\alias{qc_wrapper}
\title{Quality Control - Wrapper Function}
\usage{
qc_wrapper(
  survey_df = NULL,
  cu_df = NULL,
  wood_df = NULL,
  jam_df = NULL,
  undercut_df = NULL,
  discharge_df = NULL,
  disch_meas_df = NULL
)
}
\arguments{
\item{survey_df}{data.frame containing the survey data}

\item{cu_df}{data.frame containing the channel uni data}

\item{wood_df}{data.frame containing the wood data}

\item{jam_df}{data.frame containing the jam data}

\item{discharge_df}{data.frame containing the discharge location data}

\item{disch_meas_df}{data.frame containing the discharge measurement data}
}
\value{
a tibble with combined QC results
}
\description{
A wrapper function to perform quality control for all on-the-ground (otg)
data "types" collected using the DASH protocol.
}
\author{
Mike Ackerman
}
