% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_percents.R
\name{rescale_percents}
\alias{rescale_percents}
\title{Fix Percents Sum}
\usage{
rescale_percents(
  data_df = NULL,
  col_names = NULL,
  min_perc = 80,
  max_perc = 120
)
}
\arguments{
\item{data_df}{Input data.frame}

\item{col_names}{character vector of columns that should sum to 100}

\item{min_perc}{minimum sum of columns that should be re-scaled to 100}

\item{max_perc}{maximum sum of columns that should be re-scaled to 100}
}
\value{
a tibble with corrected percentages that sum to 100
}
\description{
This function re-scales the group of designated columns so their sum is 100\%
}
\author{
Kevin See
}
