% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_values.R
\name{rescale_values}
\alias{rescale_values}
\title{Re-scale Values}
\usage{
rescale_values(
  data_df = NULL,
  col_names = NULL,
  sum_to = 100,
  min_value = 80,
  max_value = 120
)
}
\arguments{
\item{data_df}{input data.frame}

\item{col_names}{character vector of columns that should sum to \code{sum_to}}

\item{sum_to}{the value that the columns should sum to}

\item{min_perc}{minimum sum of columns that should be re-scaled to \code{sum_to}}

\item{max_perc}{maximum sum of columns that should be re-scaled to \code{sum_to}}
}
\value{
a tibble with corrected percentages that sum to 100
}
\description{
Re-scale a group of designated columns so their sum equals \code{sum_to}
}
\author{
Kevin See and Mike Ackerman
}
