% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup_cu_undercut.R
\name{rollup_cu_undercut}
\alias{rollup_cu_undercut}
\title{Rollup Channel Unit Undercut Data}
\usage{
rollup_cu_undercut(
  undercut_df = NULL,
  fix_nas = TRUE,
  impute_cols = c("length_m", "width_25_percent_m", "width_50_percent_m",
    "width_75_percent_m"),
  ...
)
}
\arguments{
\item{undercut_df}{data.frame of \verb{otg_type =} "Undercut_4.csv" containing the individual
undercut data to be summarized (rolled up) to the channel unit scale}

\item{fix_nas}{if any of the length or width measurements for an individual undercut
is missing i.e., \code{NA}, would you like to fill them in? Default is \code{TRUE}, in which
case the \code{NA} values will be imputed using function `impute_missing_values()}

\item{impute_cols}{character vector of column names that should be imputed, if \code{fix_nas == TRUE}}

\item{...}{other arguments to \code{impute_missing_values()}}
}
\value{
a data.frame summarizing undercut data at the channel unit scale
}
\description{
Summarize individual undercut data (e.g., from \verb{otg_type =} "Undercut_4.csv) at
the channel unit scale
}
\author{
Mike Ackerman, Richie Carmichael, and Kevin See
}
