% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Undercut_4_UpperLemhi1_2019.R
\docType{data}
\name{Undercut_4_UpperLemhi1_2019}
\alias{Undercut_4_UpperLemhi1_2019}
\title{Undercut bank data from Upper Lemhi 1 survey 2019}
\format{Initially exported from Survey123 as a csv file}
\usage{
Undercut_4_UpperLemhi1_2019
}
\description{
This is an example dataset for \code{DASH()} and for undercut banks from the Upper Lemhi 1 survey in 2019.
}
\keyword{DASH}
\keyword{Lemhi}
\keyword{undercut}
