% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undercutMetrics.R
\name{undercutMetrics}
\alias{undercutMetrics}
\title{summarize the undercut bank metrics for a DASH reach}
\usage{
undercutMetrics(input = NULL)
}
\arguments{
\item{input}{the name of the tbl_df containing undercut bank data for a DASH reach.}
}
\description{
summarize the undercut bank metrics for a DASH reach
}
\author{
Richie Carmichael
}
