% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wood_2_UpperLemhi1_2019.R
\docType{data}
\name{Wood_2_UpperLemhi1_2019}
\alias{Wood_2_UpperLemhi1_2019}
\title{Large Woody Debris (LWD) data from Upper Lemhi 1 survey 2019}
\format{Initially exported from Survey123 as a csv file}
\usage{
Wood_2_UpperLemhi1_2019
}
\description{
This is an example dataset for \code{DASH()} for large woody debris from the Upper Lemhi 1 survey in 2019.
}
\keyword{DASH}
\keyword{Lemhi}
\keyword{wood}
