% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodMetrics.R
\name{woodMetrics}
\alias{woodMetrics}
\title{summarize the large woody debris metrics for a DASH reach}
\usage{
woodMetrics(input = NULL)
}
\arguments{
\item{input}{the name of the tbl_df containing large woody debris data for a DASH reach. Changing comments}
}
\description{
summarize the large woody debris metrics for a DASH reach
}
\author{
Richie Carmichael
}
