% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_otg_col_specs.R
\name{get_otg_col_specs}
\alias{get_otg_col_specs}
\title{Get/Set Column Specifications for On-The-Ground .csv Data}
\usage{
get_otg_col_specs(
  otg_type = c("surveyPoint_0.csv", "CU_1.csv", "Wood_2.csv", "Jam_3.csv",
    "Undercut_4.csv", "Discharge_5.csv", "DischargeMeasurements_6.csv")
)
}
\arguments{
\item{otg_type}{what type of on-the-ground (otg) data would the user like to deal with? As of the iniitation of the
function a character vector of the file name e.g., "surveyPoint_0.csv", "CU_1.csv", etc.}
}
\value{
an object of class "col_spec"
}
\description{
Get/set column specifications for on-the-ground (otg) data
collected using the DASH protocol.
}
\author{
Mike Ackerman
}
