% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_centerline.R
\name{qc_centerline}
\alias{qc_centerline}
\title{Quality Control - Centerlines}
\usage{
qc_centerline(
  cl_sf = NULL,
  data_id = "file_row_id",
  cols_to_check_nas = c("CU_Number", "Site_ID", "StreamName", "CU_Type", "Seg_Number",
    "Reach_Num")
)
}
\arguments{
\item{cl_sf}{The centerline \code{sf} object}

\item{cols_to_check_nas}{columns to check for NAs in}
}
\value{
a tibble with QC results
}
\description{
Quality control centerline shapefiles.
}
\author{
Kevin See
}
