% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_cu.R
\name{qc_cu}
\alias{qc_cu}
\title{Quality Control - Channel Unit Data}
\usage{
qc_cu(
  qc_df = NULL,
  cols_to_check_nas = c("Channel Unit Type", "Channel Unit Number",
    "Channel Segment Number", "Maximum Depth (m)", "ParentGlobalID")
)
}
\arguments{
\item{qc_df}{The survey data frame to be QC'd}

\item{cols_to_check_nas}{columns to check for NAs in}
}
\value{
a tibble with QC results
}
\description{
Quality control channel unit data (e.g., from CU_1.csv files) imported
using \code{read_otg_csv()} or \code{read_otg_csv_wrapper()}.
}
\author{
Mike Ackerman
}
