% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_cu_points.R
\name{qc_cu_points}
\alias{qc_cu_points}
\title{QC Channel Unit Points}
\usage{
qc_cu_points(cu_points_sf = NULL)
}
\arguments{
\item{cu_points_sf}{The sf object containing the channel unit points, perhaps
collected in the field using Field Maps}
}
\value{
a tibble with QC results
}
\description{
Quality control the channel unit points, perhaps from
Field Maps
}
\author{
Mike Ackerman
}
