% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_disch_meas.R
\name{qc_disch_meas}
\alias{qc_disch_meas}
\title{Quality Control - Discharge Measurements}
\usage{
qc_disch_meas(
  qc_df = NULL,
  cols_to_check_nas = c("Station Width", "Station Depth", "Station Velocity")
)
}
\arguments{
\item{qc_df}{The survey data frame to be QC'd}

\item{cols_to_check_nas}{columns to check for NAs in}
}
\value{
a tibble with QC results
}
\description{
Quality control discharge measurement data (e.g., from DischargeMeasurements_6.csv files) imported
using \code{read_otg_csv()} or \code{read_otg_csv_wrapper()}.
}
\author{
Kevin See
}
