% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_rollup.R
\name{qc_rollup}
\alias{qc_rollup}
\title{Quality Control - Channel Unit Rollup}
\usage{
qc_rollup(
  cu_df = NULL,
  survey_df = NULL,
  jam_df = NULL,
  undercut_df = NULL,
  wood_df = NULL,
  discharge_df = NULL,
  discharge_meas_df = NULL,
  fix_nas = TRUE,
  jam_impute_cols = c("length_m", "width_m", "height_m", "estimated_number_of_pieces"),
  undercut_impute_cols = c("length_m", "width_25_percent_m", "width_50_percent_m",
    "width_75_percent_m"),
  wood_impute_cols = c("length_m", "diameter_m"),
  ...
)
}
\arguments{
\item{cu_df}{data.frame of \verb{otg_type =} "CU_1.csv" containing data for each channel unit}

\item{survey_df}{data.frame of \verb{otg_type =} "surveyPoint_0.csv" containing information for
each site/survey}

\item{jam_df}{data.frame of \verb{otg_type =} "Jam_3.csv containing the individual
jam data to be summarized (rolled up) to the channel unit scale}

\item{undercut_df}{data.frame of \verb{otg_type =} "Undercut_4.csv" containing the individual
undercut data to be summarized (rolled up) to the channel unit scale}

\item{wood_df}{data.frame of \verb{otg_type =} "Wood_2.csv" containing the individual
wood data to be summarized by channel units}

\item{discharge_df}{data.frame of \verb{otg_type =} "Discharge_5.csv" containing the survey
and channel unit number information where discharges were measured}

\item{discharge_meas_df}{input data.frame of \verb{otg_type =} "DischargeMeasurements_6.csv"
containing the station width, depth, and velocity data to estimate discharge}

\item{fix_nas}{if any of the length, width, height, or estimated number of pieces measurements
for an individual jam is missing i.e.,\code{NA}, would you like to fill them in? Default is \code{TRUE}, in which
case the \code{NA} values will be imputed using function \code{impute_missing_values()}}

\item{...}{other arguments to \code{impute_missing_values()}}
}
\value{
a tibble with QC results
}
\description{
Quality control the channel unit rollup
performed by \code{rollup_channel_unit()}
}
\author{
Kevin See
}
