% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_survey.R
\name{qc_survey}
\alias{qc_survey}
\title{Quality Control - Survey Data}
\usage{
qc_survey(qc_df = NULL, cols_to_check_nas = c("Survey Date", "Survey Time"))
}
\arguments{
\item{qc_df}{The survey data frame to be QC'd}

\item{cols_to_check_nas}{columns to check for NAs in}
}
\value{
a tibble with QC results
}
\description{
Quality control survey data (e.g., from surveyPoint_0.csv files) imported
using \code{read_otg_csv()} or \code{read_otg_csv_wrapper()}.
}
\author{
Mike Ackerman
}
