% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_wrapper.R
\name{qc_wrapper}
\alias{qc_wrapper}
\title{Quality Control - Wrapper Function}
\usage{
qc_wrapper(
  survey_df = NULL,
  cu_df = NULL,
  wood_df = NULL,
  jam_df = NULL,
  undercut_df = NULL,
  discharge_df = NULL,
  discharge_meas_df = NULL,
  redirect_output = TRUE,
  redirect_output_path = "qc_wrapper_output.txt",
  ...
)
}
\arguments{
\item{survey_df}{data.frame containing the survey data (required to run \code{qc_wrapper()})}

\item{cu_df}{data.frame containing the channel unit data (required to run \code{qc_wrapper()})}

\item{wood_df}{data.frame containing the wood data}

\item{jam_df}{data.frame containing the jam data}

\item{discharge_df}{data.frame containing the discharge location data}

\item{discharge_meas_df}{data.frame containing the discharge measurement data}

\item{redirect_output}{would you like to redirect the output messages
from \code{qc_wrapper()} to a file instead of the R terminal? Default = \code{TRUE}.
If set to \code{FALSE}, messages will simply be written to console.}

\item{redirect_output_path}{a path and file name (e.g., .txt or .csv) to
write the \code{qc_wrapper} output messages to. Default will write
"qc_wrapper_output.txt" to the current working directory.}

\item{...}{various other inputs to \verb{qc_*} functions, if need to change from default values}
}
\value{
a tibble with combined QC results
}
\description{
A wrapper function to perform quality control for all on-the-ground (OTG)
data "types" collected using the DASH protocol.
}
\author{
Mike Ackerman
}
