% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_centerlines.R
\name{read_centerlines}
\alias{read_centerlines}
\title{Read In Centerlines}
\usage{
read_centerlines(path = ".", cl_names = "centerlines.shp")
}
\arguments{
\item{path}{A path to the directory containing the folders of interest. \code{path} should only contain
folders (no files), and each of those folders should only contain files (no folders).}

\item{cl_names}{The file name of each centerline. All centerlines should share
same name, but be saved in separate folders for each site.}
}
\value{
a shapefile of merged centerlines
}
\description{
Import centerline shapefiles (.shp). The centerlines should
contain, at the least, the variable names "site_name", "year", and "cu_num"
}
\author{
Mike Ackerman
}
