% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_otg_csv.R
\name{read_otg_csv}
\alias{read_otg_csv}
\title{Read On-The-Ground .csv Data}
\usage{
read_otg_csv(
  path = ".",
  otg_type = c("surveyPoint_0.csv", "CU_1.csv", "Wood_2.csv", "Jam_3.csv",
    "Undercut_4.csv", "Discharge_5.csv")
)
}
\arguments{
\item{path}{A path to the directory containing the folders of interest. \code{path} should only contain
folders (no files), and each of those folders should only contain files (no folders).}

\item{otg_type}{what type of on-the-ground (OTG) data would the user like to deal with? As of
the initiation of the function a character vector of the file names e.g., "surveyPoint_0.csv",
"CU_1.csv", etc.}
}
\value{
a data frame containing data of \code{otg_type}
}
\description{
Import .csv delimited files containing on-the-ground (OTG) data collected using the
DASH protocol. This function only reads in one \code{otg_type} of data at a time.
}
