% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup_cu_jam.R
\name{rollup_cu_jam}
\alias{rollup_cu_jam}
\title{Rollup Channel Unit Jam Data}
\usage{
rollup_cu_jam(jam_df = NULL, fix_nas = TRUE, ...)
}
\arguments{
\item{jam_df}{data.frame of \verb{otg_type =} "Jam_3.csv containing the individual
jam data to be summarized (rolled up) to the channel unit scale}

\item{fix_nas}{if any of the length, width, height, or estimated number of pieces measurements
for an individual jam is missing i.e.,\code{NA}, would you like to fill them in? Default is \code{TRUE}, in which
case the \code{NA} values will be imputed using function \code{impute_missing_values}.}

\item{...}{other arguments to \code{impute_missing_values}}
}
\value{
a data.frame summarizing jam at the channel unit scale
}
\description{
Summarize individual jam data (e.g., from \verb{otg_type =} "Jam_3.csv") at
the channel unit scale
}
\author{
Mike Ackerman and Richie Carmichael
}
