% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup_cu_wood.R
\name{rollup_cu_wood}
\alias{rollup_cu_wood}
\title{Rollup Channel Unit Wood Data}
\usage{
rollup_cu_wood(
  wood_df = NULL,
  fix_nas = TRUE,
  impute_cols = c("length_m", "diameter_m"),
  ...
)
}
\arguments{
\item{wood_df}{data.frame of \code{otg_type =} "Wood_2.csv" containing the individual
wood data to be summarized by channel units}

\item{fix_nas}{if any of the length or diameter measurement for individual wood pieces are missing
i.e.,\code{NA}, would you like to fill them in? Default is \code{TRUE}, in which
case the \code{NA} values will be imputed using function \code{impute_missing_values}}

\item{impute_cols}{character vector of column names that should be imputed, if \code{fix_nas == TRUE}}

\item{...}{other arguments to \code{impute_missing_values}}
}
\value{
a data.frame summarizing wood at the channel unit scale
}
\description{
Summarize individual wood piece data (e.g. from \code{otg_type =} "Wood_2.csv") at
the channel unit scale
}
\author{
Mike Ackerman, Richie Carmichael, and Kevin See
}
